+++ CHANGELOG.md
@@ -13,6 +13,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 - Contract
     - Display resident email
     - Additional agreements section
+    - Signature section
 - Validation constraint for currency including 0 €
 - Generic `TextInputField` component